design: 검색어와 일치하는 게시물이 없는 경우 화면 수정
