--- components/AllArticleList/AllArticleList.module.css
@@ -1,16 +1,0 @@
-.emptyMessageWrapper {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin-top: 1.5rem;
-}
-
-.emptyMessageText {
-  margin-top: 0.5rem;
-  margin-bottom: 3rem;
-  font-size: 1rem;
-  font-weight: normal;
-  line-height: 1.5rem;
-  color: var(--gray400);
-}
