--- pages/search/index.tsx
@@ -1,83 +1,0 @@
-import { GetServerSidePropsContext } from "next";
-import { useEffect, useState } from "react";
-import { Article } from "@/types/article";
-import axios from "@/lib/axios";
-
-import styles from "@/styles/boards.module.css";
-
-import LinkButton from "@/components/LinkButton";
-import SearchForm from "@/components/SearchForm";
-import Sort from "@/components/Sort";
-import BestArticleList from "@/components/BestArticleList/BestArticleList";
-import AllArticleList from "@/components/AllArticleList/AllArticleList";
-
-export async function getServerSideProps(context: GetServerSidePropsContext) {
-  const { keyword } = context.query;
-  const response = await axios.get(`/articles?keyword=${keyword}`);
-  const articles = response.data.list ?? [];
-
-  return {
-    props: { articles },
-  };
-}
-
-interface BoardProps {
-  articles: Article[];
-}
-
-function Board({ articles }: BoardProps) {
-  // const [bestArticles, setBestArticles] = useState<Article[]>([]);
-  // const [pageSize, setPageSize] = useState<number>(3);
-
-  // useEffect(() => {
-  //   const updatePageSize = () => {
-  //     // 서버 사이드에서의 window 접근 방지
-  //     if (typeof window !== "undefined") {
-  //       const { innerWidth } = window;
-  //       if (innerWidth < 377) setPageSize(1);
-  //       else if (innerWidth < 745) setPageSize(2);
-  //       else setPageSize(3);
-  //     }
-  //   };
-  //   window.addEventListener("resize", updatePageSize);
-
-  //   return () => {
-  //     window.removeEventListener("resize", updatePageSize);
-  //   };
-  // }, []);
-
-  // async function getBestArticles(pageSize: number) {
-  //   const response = await axios.get(
-  //     `/articles/?orderBy=like&pageSize=${pageSize}`
-  //   );
-  //   setBestArticles(response.data.list ?? []);
-  // }
-
-  // useEffect(() => {
-  //   getBestArticles(pageSize);
-  // }, [pageSize]);
-
-  return (
-    <main>
-      {/* <section className={styles.section}>
-        <h3 className={`${styles.sectionTitle} ${styles.best}`}>
-          베스트 게시글
-        </h3>
-        <BestArticleList articles={bestArticles} />
-      </section> */}
-      <section>
-        <div className={styles.titleWrapper}>
-          <h3 className={styles.sectionTitle}>게시글</h3>
-          <LinkButton href="" text="글쓰기" />
-        </div>
-        <div className={styles.filterWrapper}>
-          <SearchForm />
-          {/* <Sort setArticles={setArticles} /> */}
-        </div>
-        <AllArticleList articles={articles} />
-      </section>
-    </main>
-  );
-}
-
-export default Board;
